# src/resources/openapi/swagger_apartment_management.yaml
openapi: 3.0.3
info:
  title: Swagger Apartment Management System - OpenAPI 3.0
  description: |-
    This is a sample Apartment Management System Server based on the OpenAPI 3.0 specification.
    Some useful links:
    - [The Apartment Management System repository](https://github.com/swagger-api/swagger-petstore)
  termsOfService: http://swagger.io/terms/
  contact:
    email: 283531@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://pwrapartmentmanagement/api/v3
...
paths:
...
  /apartment/task/create:
    post:
      tags:
        - task
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDTO'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuccessApiResponse'
        '400':
          description: Bad Request - Invalid user ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid user ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The user could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The user could not be found
  /apartment/task/{taskID}/update:
    post:
      tags:
        - task
      summary: Update a task
      operationId: updateTask
      parameters:
        - name: taskID
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDTO'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuccessApiResponse'
        '400':
          description: Bad Request - Invalid user ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid user ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The user could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The user could not be found
  /apartment/task/{taskID}:
    get:
      tags:
        - task
      summary: Retrieve a specific task by ID
      operationId: getTaskById
      parameters:
        - name: taskID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved the task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '400':
          description: Bad Request - Invalid user ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid user ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The user could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The user could not be found
  /apartment/task:
    get:
      tags:
        - task
      summary: Retrieve all tasks
      operationId: getAllTasks
      responses:
        '200':
          description: Successfully retrieved all tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TasksResponse'
        '400':
          description: Bad Request - Invalid user ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid user ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The user could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The user could not be found
...
  /apartment/task/{taskID}/delete:
    delete:
      tags:
        - task
      summary: Delete a maintenance task
      operationId: deleteMaintenanceTask
      parameters:
        - name: taskID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Maintenance task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessApiResponse'
        '400':
          description: Bad Request - Invalid user ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 400
                message: Bad Request - Invalid user ID supplied
        '401':
          description: Unauthorized - Authentication credentials were missing or incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 401
                message: Unauthorized - Authentication credentials were missing or incorrect
        '403':
          description: Forbidden - The request was valid but you do not have the necessary permissions for the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                message: Forbidden - The request was valid but you do not have the necessary permissions for the resource
        '404':
          description: Not Found - The user could not be found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 404
                message: Not Found - The user could not be found
...
components:
  schemas:
    SuccessApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          default: 201
        message:
          type: string
          default: "Success."
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    TaskDTO:
      type: object
      properties:
        taskName:
          type: string
          example: "Fix light fixture"
        userID:
          type: integer
          example: 1
        taskDescription:
          type: string
          example: "The light fixture in the living room is flickering."
        assignedName:
          type: string
          example: "John Doe"
      required:
        - taskName
        - userID
        - taskDescription
        - assignedName
...
    TasksResponse:
      type: array
      items:
        $ref: '#/components/schemas/TaskDTO'
...